fix: rsbuild doctor add e2e tests
fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests

fix: rsbuild doctor add e2e tests
