+++ .github/workflows/test-ubuntu-doctor.yml
@@ -1,0 +1,59 @@
+name: Doctor E2E (Ubuntu)
+
+# Controls when the action will run.
+on:
+  # Triggers the workflow on pull request events but only for the main branch
+  pull_request:
+    branches: [main]
+
+  push:
+    branches: [main]
+    paths:
+      - "packages/doctor-*"
+
+  merge_group:
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+  # ===== e2e for doctor =======
+    e2e-ubuntu-doctor:
+      runs-on: ubuntu-latest
+      strategy:
+        matrix:
+          # Rsbuild need to compat Node.js 16
+          node-version: [16.x] 
+
+      # Steps represent a sequence of tasks that will be executed as part of the job
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v3
+          with:
+            fetch-depth: 10
+
+        - name: Install Pnpm
+          run: corepack enable
+
+        - name: Check skip CI
+          run: echo "RESULT=$(node ./scripts/skipCI.js)" >> "$GITHUB_OUTPUT"
+          id: skip-ci
+
+        - name: Log skip CI result
+          run: echo "${{steps.skip-ci.outputs.RESULT}}"
+
+        - name: Setup Node.js ${{ matrix.node-version }}
+          if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
+          uses: actions/setup-node@v3
+          with:
+            node-version: ${{ matrix.node-version }}
+            cache: 'pnpm'
+
+        - name: Install Dependencies
+          if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
+          run: pnpm install && cd ./e2e && npx playwright install
+
+        - name: E2E Test
+          if: ${{steps.skip-ci.outputs.RESULT != 'true'}}
+          run: pnpm run e2e:doctor
