+++ e2e/cases/doctor-rspack/tests/fixtures/a.js
