+++ e2e/cases/doctor-rspack/tests/fixtures/b.js
