+++ e2e/cases/doctor-rspack/tests/fixtures/loaders/comment.js
