+++ e2e/cases/doctor-rspack/tests/fixtures/loaders/serialize-query-to-comment.js
