+++ e2e/cases/doctor-rspack/tests/test-utils.ts
@@ -1,9 +1,9 @@
 import { DoctorRspackPluginOptions } from '@rsbuild/doctor-core/types';
+import { RsbuildDoctorRspackPlugin } from '@rsbuild/doctor-rspack-plugin';
 import { Linter } from '@rsbuild/doctor-types';
 import { File } from '@rsbuild/doctor-utils/build';
 import { tmpdir } from 'os';
 import path from 'path';
-import { RsbuildDoctorRspackPlugin } from '../src';
 
 export function createDoctorPlugin<T extends Linter.ExtendRuleData[]>(
   options: DoctorRspackPluginOptions<T>,
@@ -20,10 +20,14 @@ export function createDoctorPlugin<T extends Linter.ExtendRuleData[]>(
     tmpdir(),
     `./${Date.now()}/rsbuild_doctor_rspack_plugin_test`,
   );
-  plugin.sdk.setOutputDir(outdir);
 
   plugin.sdk.hooks.afterSaveManifest.tapPromise('REMOVE_TMP_DIR', async () => {
-    await File.fse.remove(plugin.sdk.outputDir);
+    plugin.sdk.setOutputDir(outdir);
+    try {
+      await File.fse.remove(plugin.sdk.outputDir);
+    } catch (e) {
+      console.error(e);
+    }
   });
 
   return plugin;
