+++ e2e/cases/doctor-webpack/fixtures/a.js
