+++ e2e/cases/doctor-webpack/fixtures/b.js
