+++ e2e/cases/doctor-webpack/fixtures/loaders/comment.js
