+++ e2e/cases/doctor-webpack/fixtures/loaders/serialize-query-to-comment.js
