+++ e2e/package.json
@@ -6,7 +6,8 @@
     "test": "pnpm run test:rspack && pnpm run test:webpack",
     "test:webpack": "PROVIDE_TYPE=webpack playwright test",
     "test:rspack": "playwright test",
-    "test:webpack-swc": "PROVIDE_TYPE=webpack playwright test swc"
+    "test:webpack-swc": "PROVIDE_TYPE=webpack playwright test swc",
+    "test:doctor": "playwright test -c playwright.config.doctor.ts"
   },
   "dependencies": {
     "devcert": "1.2.2",
@@ -19,10 +20,16 @@
   "devDependencies": {
     "@playwright/test": "1.33.0",
     "@rsbuild/core": "workspace:*",
+    "@rsbuild/doctor-core": "workspace:*",
+    "@rsbuild/doctor-rspack-plugin": "workspace:*",
+    "@rsbuild/doctor-sdk": "workspace:*",
+    "@rsbuild/doctor-types": "workspace:*",
+    "@rsbuild/doctor-utils": "workspace:*",
+    "@rsbuild/doctor-webpack-plugin": "workspace:*",
     "@rsbuild/plugin-assets-retry": "workspace:*",
     "@rsbuild/plugin-babel": "workspace:*",
-    "@rsbuild/plugin-css-minimizer": "workspace:*",
     "@rsbuild/plugin-check-syntax": "workspace:*",
+    "@rsbuild/plugin-css-minimizer": "workspace:*",
     "@rsbuild/plugin-image-compress": "workspace:*",
     "@rsbuild/plugin-node-polyfill": "workspace:*",
     "@rsbuild/plugin-pug": "workspace:*",
@@ -40,11 +47,14 @@
     "@rsbuild/plugin-vue2": "workspace:*",
     "@rsbuild/plugin-vue2-jsx": "workspace:*",
     "@rsbuild/shared": "workspace:*",
+    "@rsbuild/test-helper": "workspace:*",
     "@rsbuild/webpack": "workspace:*",
+    "@rspack/core": "0.3.14",
     "@types/lodash": "^4.14.200",
     "@types/node": "^16",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/webpack": "5.28.0",
     "fast-glob": "^3.3.1",
     "playwright": "1.33.0",
     "typescript": "^5.3.0"
