+++ e2e/playwright.config.doctor.ts
@@ -1,0 +1,5 @@
+import { defineConfig } from '@playwright/test';
+
+export default defineConfig({
+  testMatch: ['**/cases/doctor-*/**/**.test.ts'],
+});
