+++ packages/doctor-rspack-plugin/package.json
@@ -15,8 +15,7 @@
   "scripts": {
     "dev": "npm run start",
     "start": "modern build -w",
-    "build": "modern build",
-    "test": "vitest run"
+    "build": "modern build"
   },
   "dependencies": {
     "@rsbuild/doctor-core": "workspace:*",
